refactor(llm): split chat-executor.ts into focused modules
Decompose the 5,741-line ChatExecutor monolith into 8 responsibility-focused
files while preserving all public API surfaces via re-exports.

New modules:
- chat-executor-types.ts (620 lines): all types, interfaces, error class
- chat-executor-constants.ts (201 lines): constants and tool classification sets
- chat-executor-tool-utils.ts (131 lines): standalone tool helper functions
- chat-executor-text.ts (658 lines): text processing and formatting
- chat-executor-recovery.ts (170 lines): semantic keys, recovery hints
- chat-executor-planner.ts (1,155 lines): planner parsing/validation/building
- chat-executor-verifier.ts (421 lines): subagent verifier functions

All private static methods converted to standalone exported functions.
Zero breaking changes — llm/index.ts and all 17 consumer files unchanged.
